requirements.txt
# Core dependencies
streamlit==1.31.1
python-dotenv==1.0.1
pyyaml==6.0.1

# LangChain and LLMs
langchain==0.1.10
langchain-openai==0.0.8
langchain-community==0.0.24

# Vector stores
chromadb==0.4.22

# Document parsing
pypdf==4.0.2
python-docx==1.1.0
pandas==2.2.0

# Optional: Ollama support
# ollama==0.1.6

# Testing
pytest==8.0.1
pytest-cov==4.1.0
